honeypot-backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
docker>=6.0.1
pydantic>=1.10.7
python-dotenv>=1.0.0
python-multipart>=0.0.6
websockets>=11.0.2